requirements.txt
django
djangorestframework
google-genai
django-cors-headers
python-dotenv
langchain 
langchain_groq 
groq
google-cloud-storage
googlemaps
bs4
requests
PyPDF2
eventregistry